requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
python-multipart
pymupdf
pytesseract
transformers==4.44.0
torch==2.4.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
yake
reportlab
aiofiles
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
python-decouple
google-auth
